Introduce test for macro
