+++ sewup-derive/Cargo.toml
@@ -14,9 +14,12 @@ proc-macro = true
 [dependencies]
 regex = "1"
 tiny-keccak = { version = "2.0.2", features = ["keccak"] }
+proc-macro-error = "1.0.4"
 
 [dev-dependencies]
 hex-literal = "0.3.1"
+trybuild = "1"
+rustversion = "1"
 
 [features]
 kv = []
