+++ sewup-derive/tests/compile-fail/ewasm-fn-sig.rs
@@ -1,0 +1,5 @@
+use sewup_derive::ewasm_fn_sig;
+
+fn main() {
+    let a = ewasm_fn_sig!(some);
+}
