+++ sewup-derive/tests/compile-fail/ewasm-fn-sig.stderr
@@ -1,0 +1,7 @@
+error[E0425]: cannot find value `SOME_SIG` in this scope
+ --> $DIR/ewasm-fn-sig.rs:4:13
+  |
+4 |     let a = ewasm_fn_sig!(some);
+  |             ^^^^^^^^^^^^^^^^^^^ not found in this scope
+  |
+  = note: this error originates in a macro (in Nightly builds, run with -Z macro-backtrace for more info)
