+++ sewup-derive/tests/ui.rs
@@ -1,0 +1,9 @@
+#[rustversion::stable]
+#[test]
+fn ui() {
+    use trybuild::TestCases;
+
+    let t = TestCases::new();
+    t.compile_fail("tests/compile-fail/*.rs");
+    t.pass("tests/compile-pass/*.rs");
+}
